+++ packages/clerk_flutter/lib/src/utils/localization_extensions.dart
@@ -14,10 +14,10 @@ extension ClerkAuthErrorExtension on clerk.ClerkError {
       clerk.ClerkErrorCode.cannotDeleteSelf => l10ns.cannotDeleteSelf,
       clerk.ClerkErrorCode.jwtPoorlyFormatted =>
         l10ns.jwtPoorlyFormatted(argument.toString()),
-      clerk.ClerkErrorCode.noAssociatedStrategy =>
-        l10ns.noAssociatedStrategy(argument.toString()),
       clerk.ClerkErrorCode.noAssociatedCodeRetrievalMethod =>
         l10ns.noAssociatedCodeRetrievalMethod(argument.toString()),
+      clerk.ClerkErrorCode.noAssociatedStrategy =>
+        l10ns.noAssociatedStrategy(argument.toString()),
       clerk.ClerkErrorCode.noSessionFoundForUser =>
         l10ns.noSessionFoundForUser(argument.toString()),
       clerk.ClerkErrorCode.noSessionTokenRetrieved =>
@@ -28,11 +28,14 @@ extension ClerkAuthErrorExtension on clerk.ClerkError {
         l10ns.noSuchFirstFactorStrategy(argument.toString()),
       clerk.ClerkErrorCode.noSuchSecondFactorStrategy =>
         l10ns.noSuchSecondFactorStrategy(argument.toString()),
+      clerk.ClerkErrorCode.noUserAttributeForField =>
+        l10ns.noUserAttributeForField(argument.toString()),
       clerk.ClerkErrorCode.passwordMatchError => l10ns.passwordMatchError,
       clerk.ClerkErrorCode.passwordResetStrategyError =>
         l10ns.unsupportedPasswordResetStrategy(argument.toString()),
       clerk.ClerkErrorCode.serverErrorResponse =>
         l10ns.serverErrorResponse(argument.toString()),
+      clerk.ClerkErrorCode.tooManyRetries => l10ns.tooManyRetries,
       clerk.ClerkErrorCode.unknownError =>
         l10ns.unknownError(argument.toString()),